+++ packages/code-editor/env.d.ts
@@ -1,0 +1,6 @@
+declare module 'monaco-vim' {
+  function initVimMode(
+    editor: ReturnType<(typeof import('monaco-editor'))['editor']['create']>,
+    statusBar: HTMLElement
+  ): { dispose: () => void };
+}